client: extract formatter pkg
