+++ tsuru/formatter/date.go
@@ -1,19 +1,19 @@
-// Copyright 2017 tsuru-client authors. All rights reserved.
+// Copyright 2018 tsuru-client authors. All rights reserved.
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package client
+package formatter
 
 import (
 	"fmt"
 	"time"
 )
 
-func formatDate(date time.Time) string {
+func FormatDate(date time.Time) string {
 	return date.Local().Format(time.Stamp)
 }
 
-func formatDuration(duration *time.Duration) string {
+func FormatDuration(duration *time.Duration) string {
 	if duration == nil {
 		return "…"
 	}
@@ -24,6 +24,6 @@ func formatDuration(duration *time.Duration) string {
 	return fmt.Sprintf("%02d:%02d", minutes, seconds)
 }
 
-func formatDateAndDuration(date time.Time, duration *time.Duration) string {
-	return fmt.Sprintf("%s (%s)", formatDate(date), formatDuration(duration))
+func FormatDateAndDuration(date time.Time, duration *time.Duration) string {
+	return fmt.Sprintf("%s (%s)", FormatDate(date), FormatDuration(duration))
 }
