+++ tsuru/formatter/suite_test.go
@@ -1,0 +1,17 @@
+// Copyright 2018 tsuru-client authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+package formatter
+
+import (
+	"testing"
+
+	"gopkg.in/check.v1"
+)
+
+type S struct{}
+
+var _ = check.Suite(&S{})
+
+func Test(t *testing.T) { check.TestingT(t) }
